RUMM-209 Add user agent into traces intake request
