+++ dd-sdk-android/src/main/kotlin/com/datadog/android/log/internal/net/LogsOkHttpUploader.kt
@@ -1,7 +1,5 @@
 package com.datadog.android.log.internal.net
 
-import android.os.Build
-import com.datadog.android.BuildConfig
 import com.datadog.android.core.internal.net.DataOkHttpUploader
 import java.util.Locale
 import okhttp3.OkHttpClient
@@ -14,35 +12,13 @@ internal open class LogsOkHttpUploader(
 
     // region DataOkHttpUploader
 
-    override fun headers(): MutableMap<String, String> {
-        val headers = super.headers()
-        headers[HEADER_UA] = userAgent
-        return headers
-    }
-
     override fun setEndpoint(endpoint: String) {
         super.setEndpoint(buildUrl(endpoint, token))
     }
 
     // endregion
 
-    // region Internal
-
-    private val userAgent = System.getProperty(SYSTEM_UA).let {
-        if (it.isNullOrBlank()) {
-            "Datadog/${BuildConfig.VERSION_NAME} " +
-                "(Linux; U; Android ${Build.VERSION.RELEASE}; " +
-                "${Build.MODEL} Build/${Build.ID})"
-        } else {
-            it
-        }
-    }
-
-    // endregion
-
     companion object {
-        private const val HEADER_UA = "User-Agent"
-        const val SYSTEM_UA = "http.agent"
         private const val QP_SOURCE = "ddsource"
         private const val DD_SOURCE_MOBILE = "mobile"
         internal const val UPLOAD_URL =
