+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/net/LogsOkHttpUploaderTest.kt
@@ -1,13 +1,9 @@
 package com.datadog.android.log.net
 
-import android.os.Build
-import com.datadog.android.BuildConfig
 import com.datadog.android.core.internal.net.DataOkHttpUploaderTest
 import com.datadog.android.log.internal.net.LogsOkHttpUploader
 import java.util.concurrent.TimeUnit
 import okhttp3.OkHttpClient
-import okhttp3.mockwebserver.RecordedRequest
-import org.assertj.core.api.Assertions.assertThat
 
 internal class LogsOkHttpUploaderTest : DataOkHttpUploaderTest<LogsOkHttpUploader>() {
 
@@ -21,19 +17,6 @@ internal class LogsOkHttpUploaderTest : DataOkHttpUploaderTest<LogsOkHttpUploade
                 .writeTimeout(TIMEOUT_TEST_MS, TimeUnit.MILLISECONDS)
                 .build()
         )
-    }
-
-    override fun assertHeaders(request: RecordedRequest) {
-        super.assertHeaders(request)
-        val expectedUserAgent = if (fakeUserAgent.isBlank()) {
-            "Datadog/${BuildConfig.VERSION_NAME} " +
-                    "(Linux; U; Android ${Build.VERSION.RELEASE}; " +
-                    "${Build.MODEL} Build/${Build.ID})"
-        } else {
-            fakeUserAgent
-        }
-        assertThat(request.getHeader("User-Agent"))
-            .isEqualTo(expectedUserAgent)
     }
 
     override fun urlFormat(): String {
